% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allHex}
\alias{allHex}
\title{Random set of colors for the plotting of biogegoraphic partitionings}
\format{
An object of class \code{character} of length 3544.
}
\usage{
data(allHex)
}
\description{
The  first 10-15 colors are manually selected for optimal contrast, the rest are procedurally generated.
}
\keyword{datasets}
