% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bgpart}
\alias{bgpart}
\title{Biogeographic partitioning of occurrence data.}
\usage{
bgpart(dat, tax, cell, bin = NULL, ocq = 0, base = "network",
  feedback = FALSE, slq = 50, tracing = FALSE, coll = NULL,
  cols = TRUE, omitted = TRUE, ...)
}
\arguments{
\item{dat}{\code{(data.frame)} Occurrence table.}

\item{tax}{\code{(character)} Variable name of the occurring taxa (variable type: \code{factor} or \code{character} - such as \code{"genus"}}

\item{cell}{(\code{character}) The variable name of the cell, site or sample entries.}

\item{bin}{\code{(character)} Variable name of the time bin numbers of the occurrences. This variable should be \code{numeric}.}

\item{ocq}{\code{numeric} The minimum occurrence/cell quota. Cells with a lower number of occurrences will be omitted from the analysis.}

\item{base}{(\code{character}) The basic type of the partitiong method. Currently only \code{"network"} or \code{"distance"}-based methods are provided, 
which are implemented in the \code{\link{groupNet}} and \code{\link{groupDist}} functions, respectively.}

\item{feedback}{(\code{logical}) Should the functions output information about progress?}

\item{slq}{\code{numeric} Only used if bin is not NULL. The minimum occurrence/time slice quota. Time slices with a lower number of occurrences will be omitted from the analysis.}

\item{coll}{Used only for the 'network' approaches implemented in \code{\link{groupNet}}, and when bipartite is FALSE. The column name of the collection/sample identifiers. Setting this value to a valid column name will enforce a collection-based correction (Rojas et al., 2016), and will make 'sampcorr' argument irrelevant.}

\item{cols}{\code{logical} Should colors be automatically assigned to each group?}

\item{...}{Arguments passed to the \code{\link{groupNet}} or \code{\link{groupDist}} functions.}
}
\description{
This function will produce bioregion/cell membership tables based on a variety of partitioning algorithms.
}
\details{
A lot stuff should come here!
}
