% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_related.R
\name{bidensity}
\alias{bidensity}
\title{Rescaled bipartite density (BC) of Sidor et al. 2013}
\usage{
bidensity(dat, tax, loc, bin = NULL)
}
\arguments{
\item{dat}{(\code{data.frame}) Occurrence dataset.}

\item{tax}{(\code{character}) Column name of taxon entries}

\item{loc}{(\code{character}) Column name of locality/cell entries.}

\item{bin}{(\code{character}) Column name of bin ids. Optional, if the function is to be iterated for multiple bins.}
}
\description{
The proportion of manifested connections out of all potential connections.
}
\details{
The variables is called BC in Sidor et al. 2013 and is caclulated as:
BC = (O-N)/LN-N'

"where O is the number of links in the occurrence network (number of occurrences),
 N is the number of taxa, and L is the number of localities. The numerator is the number of occurrences of taxa beyond a single locality (hence why N is subtracted from O), and
the denominator is the number of occurrences that could conceivably
occur (LN), minus one occurrence for each species because
each species must occur at least once. This measure is
bounded between 0 (when O = N) and 1 (when O = LN), which
correspond to extreme occurrence network topologies of minimum
and maximum homogeneity"

Sidor, C. A., D. A. Vilhena, K. D. Angielczyk, A. K. Huttenlocker, S. J. Nesbitt, B. R. Peecook, J. S. Steyer, R. M. H. Smith, and L. A. Tsuji. 2013: Provincialization of terrestrial faunas following the end-Permian mass extinction. Proceedings of the National Academy of Sciences 110:8129-8133.
}
\examples{
data(ceno6)
bidensity(ceno6, tax="trinomen", loc="icos", bin="stg")
}
