% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemism.R
\name{endemism}
\alias{endemism}
\title{Endemism of a set of localities}
\usage{
endemism(dat, tax = NULL, loc = NULL, prop = TRUE)
}
\arguments{
\item{dat}{\code{(data.frame)} Occurrence table, continency (rows are taxa, columns are localities) table or a RasterStack.}

\item{tax}{\code{(character)} Variable name of the occurring taxa (variable type: \code{factor} or \code{character} - such as \code{"genus"}}

\item{loc}{\code{(character)} Name of the locality variable, for instance samples IDs, sites, regions or geographic cells.}

\item{prop}{\code{(logical)} Set to \code{TRUE} for proportional endemism (default). Setting this to \code{FALSE} will output the the number of endemic species.}
}
\description{
The function calculates either the proportion of endemic taxa at every locality or counts the number of endemic species
}
