% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkBased.R
\name{exportPajekUndirected}
\alias{exportPajekUndirected}
\title{Export a unipartite or bipartite graph to a pajek file}
\usage{
exportPajekUndirected(graph, file)
}
\arguments{
\item{graph}{An igraph graph.}

\item{file}{The file path of the output file.}
}
\value{
The function returns no value.
}
\description{
Export a unipartite or bipartite graph to a pajek file
}
