% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceBased.R
\name{groupDist}
\alias{groupDist}
\title{Single time slice compositional partitioning based on distance-based cluster analysis}
\usage{
groupDist(
  contingency,
  binary = TRUE,
  dist = "jaccard",
  method = "ward.D2",
  counts = FALSE,
  kgs = FALSE,
  plot = FALSE,
  feedback = FALSE,
  ...
)
}
\arguments{
\item{contingency}{(\code{matrix}) Numeric matrix, species in rows, localities in columns.}

\item{binary}{(\code{logical}) If set to \code{TRUE}, the contingency table will be reduced to a binary matrix before the distances are calculated.}

\item{dist}{(\code{character}) The distance calculation method, the \code{method} argument of the \code{\link[vegan]{vegdist}} function in vegan.}

\item{method}{(\code{character}) The type of clustering used, the \code{method} argument of \code{\link[stats]{hclust}} function.}

\item{counts}{(\code{logical}) If \code{binary} is \code{FALSE}, should proportions be used, or counts?}

\item{kgs}{(\code{logical}) Should the \code{\link[maptree]{kgs}} (\code{TRUE}) or the \code{\link[stats]{cutree}} (\code{FALSE}) function be used for cutting the trees?}

\item{plot}{(\code{logical}) Should the dendrogram be plotted?}

\item{...}{functions passed to the \code{\link[stats]{cutree}} or the \code{\link[maptree]{kgs}} functions.}
}
\description{
This function calculates a distance matrix from the data and runs cluster analysis on it. The dendrogram is then cut to form higher level units.
}
