% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkBased.R
\name{groupNet}
\alias{groupNet}
\title{Single time slice,network-based compositional (e.g. biogeographic) partitioning}
\usage{
groupNet(
  contingency,
  bipartite = FALSE,
  method = "infomap",
  export = NULL,
  console = FALSE,
  sampcorr = NULL,
  feedback = FALSE,
  onlyloc = TRUE,
  ...
)
}
\arguments{
\item{contingency}{(\code{matrix}) Numeric matrix, species in rows, localities in columns.}

\item{bipartite}{\code{logical} Setting this argument to \code{FALSE} will project the bipartite graph calculated from the contingency matrix, and make a unipartite grpah of the localities.}

\item{method}{Community clustering method. "infomap" "louvain" "netcarto" NULL: network is output}

\item{export}{Path to save the graph in rgexf format}

\item{console}{\code{logical} Argument specifying whether an igraph, or a console application should be used. Currently only implemented for 'infomap'.}

\item{sampcorr}{Can be either "occ", "dom" or a named numeric vector, where the names are the column names of \code{contingency}.}

\item{feedback}{(\code{logical}) Should the functions output information about progress?}

\item{onlyloc}{\code{logical} Argumnet valid only if bipartite=TRUE. Should the taxa be reported too?}
}
\description{
This function runs network-based compositional partitioning on a contingency matrix.
}
