% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Z3 resolution land polygons from an older version of Openstreetmap}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1006 rows and 1 columns.
}
\usage{
data(land)
}
\description{
Land polygons useful for plotting of global maps.
}
\keyword{datasets}
