% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{rangerarity}
\alias{rangerarity}
\title{Range rarity metric of endemism}
\usage{
rangerarity(dat, tax, loc, fine = NULL)
}
\arguments{
\item{dat}{\code{data.frame or RasterStack} The occurrence table.}

\item{tax}{\code{character} Variable name of the taxon names.}

\item{loc}{\code{character} Variable name of the locality information (e.g. geographic cells). Range rarity will be output for every entry in this column.}

\item{fine}{\code{character} Variable name of the basic geographic information (optional) that is clustered to coarser units (in \code{loc}).}
}
\description{
The function calculates the cumulated range rarity for every geographic cell given an occurrence table
}
\details{
The metric was originally impelmented by Selig et al., 2014 (Plos ONE). For every cell or locality (\code{loc} argument), 
the function sums up the inverse occupancy (1/occupancy) of every taxon that occurrs there. 
If available, the function will weigh each of these taxon/locality specific values by a fraction. 
As in Selig et al. (2014), this is the proportion of finer-scale occurrence-points (e.g. coordinates) in the locality (cell) compared to the maximum observed number of fine-scale occurrences in that particular locality (cell). 
In other words, entries in \code{loc} cluster entries in \code{fine}. The weights will be the number a taxon has compared to the maximum number observed in the whole dataset.
}
