% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_related.R
\name{submodularity}
\alias{submodularity}
\title{Modularity in time-slice specific graph subsets}
\usage{
submodularity(x, graph, bin = NULL)
}
\arguments{
\item{x}{(\code{data.frame}) The output of the \code{\link{bgpart}} function.}

\item{graph}{(\code{igraph}) The occurrence graph, which can also be retrieved by \code{\link{bgpart}}.}

\item{bin}{(\code{character}) The column name of the bin identifier.}
}
\description{
The function returns the modularity in an occurrence graph given a provided partitioning
}
\examples{
data(ceno6)
# 1. sinlge-slice partitioning
oneC6 <- bgpart(ceno6,bin=NULL, tax="trinomen", cell="icos", ocq=10, base="network", method="infomap")
# return the used graph
oneC6_graph <- bgpart(ceno6,bin=NULL, tax="trinomen", cell="icos", ocq=10, base="network", method=NULL)
# modularity in this single slice
submodularity(oneC6, oneC6_graph)

# 2. multi-slice partitioning
trace <- bgpart(ceno6,bin="stg", tax="trinomen", cell="icos", ocq=10, base="network", method="infomap", tracing=TRUE)
# return the used graph
trace_graph <- bgpart(ceno6,bin="stg", tax="trinomen", cell="icos", ocq=10, base="network", method=NULL, tracing=TRUE)
# modularity in this single slice
submodularity(trace, trace_graph, bin="stg")
}
