% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{bgstats}
\alias{bgstats}
\title{Statistics calculated from a biogeographic partitioning}
\usage{
bgstats(
  bg,
  cell,
  bin = NULL,
  bu = "grouping",
  mjc = 3,
  bb = TRUE,
  noNAStart = FALSE,
  propsing = FALSE
)
}
\arguments{
\item{bg}{Output of the \code{\link{bgpart}} function, with \code{tracing=TRUE}.}

\item{cell}{(\code{character}) Variable name of the spatial cells.}

\item{bin}{(\code{character}) Variable name of the time slice identifier.}

\item{bu}{(\code{character}) Variable name of grouping (biogeographic units).}

\item{mjc}{(\code{numeric}) Minimum number of jointly sampled cells betwen a pair of time slices. For slicewise results, this can be set to \code{NULL}, which will not return patterns of jointly sampled cells.}

\item{bb}{\code{logical} Should by-bioregion turnover be calculated (requires the divDyn package).}

\item{noNAStart}{(\code{logical}) Same as divDyn.}

\item{propSing}{(\code{logical}) Should single-interval regions be counted in the emergence/disintegration proportions? Default to \code{FALSE} as in Kocsis et al. 2018, Proceedings B.}
}
\description{
This function will output various statistics of biogeographic partitioning files
}
\details{
Description of variables.
}
